phase2a/pyproject.toml
[tool.poetry]
name = "parscale-ear"
version = "0.3.0-lite-hybrid"
description = "ParScale-EAR: Lite-Hybrid Energy Autoregressive Model with dual-branch processing"
authors = ["ParScale Team <parscale-team@example.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/parscale/parscale-ear"
homepage = "https://github.com/parscale/parscale-ear"
documentation = "https://github.com/parscale/parscale-ear/blob/main/README.md"
keywords = ["machine-learning", "computer-vision", "image-generation", "cuda", "pytorch"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research", 
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = "^3.10"
torch = {version = "2.0.1+cu118", source = "pytorch-cu118"}
torchvision = {version = "0.15.2+cu118", source = "pytorch-cu118"}
torchaudio = {version = "2.0.2+cu118", source = "pytorch-cu118"}
pytorch-fid = "0.3.0"
numpy = "^1.26.0"
tqdm = "^4.66.0"
pillow = "^10.0.0"
scipy = "^1.11.0"
matplotlib = "^3.7.0"
pytest = "^7.4.0"
pytest-xdist = "^3.3.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"

[[tool.poetry.source]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["ci/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--verbose --tb=short --maxfail=1"